+++ app/contexts/SendTokenContext.tsx
@@ -27,7 +27,7 @@ import { useSend } from 'screens/send/useSend'
 import { TokenType, TokenWithBalance } from 'store/balance'
 import { TokenSymbol } from 'store/network'
 import { useSelector } from 'react-redux'
-import { selectActiveAccount } from 'store/accounts'
+import { selectActiveAccount } from 'store/account'
 import sendService from 'services/send/SendService'
 import { SendState } from 'services/send/types'
 import { BigNumber } from 'ethers'
